Moved plot related files to their new expected location
