+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/plot/spectrogram/SpectrogramPlotViewModel.kt
@@ -14,7 +14,6 @@ import kotlinx.coroutines.withContext
 import org.noiseplanet.noisecapture.audio.ANDROID_GAIN
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.measurements.MeasurementsService
-import org.noiseplanet.noisecapture.ui.components.measurement.SpectrogramBitmap
 
 class SpectrogramPlotViewModel(
     private val measurementsService: MeasurementsService,
